Refactor Component class name and file names
