+++ recsa/classes/component/__init__.py
@@ -1,2 +1,2 @@
 from .bindsite_existence_check import *
-from .component_structure import *
+from .component import *
